+++ CHANGELOG.md
@@ -6,6 +6,8 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 ## [Unreleased]
 
+* Added support for UpdateAndFetchResults for `SyncConnectionWrapper`, allowing to use `save_changes`
+
 ## [0.7.4] - 2025-11-07
 
 * Fixed an issue with dropping uncached mysql statements